# docs/openapi.yaml
openapi: 3.1.1
info:
  title: MiniBlog API - Proyecto Integrador M2
  description: API REST para gestionar autores, publicaciones y comentarios literarios.
  version: 2.0.0
  contact:
    name: Fernanda Posada García
    email: ferposadagarcia@gmail.com
servers:
  - url: http://localhost:3000
    description: Servidor de Desarrollo
  - url: https://miniblogapi-production-00c8.up.railway.app
    description: Servidor de Producción en Railway

paths:
  /health:
    get:
      summary: Estado de la API
      responses:
        '200':
          description: API funcionando.
  /authors:
    get:
      summary: Listar todos los autores
      responses:
        '200':
          description: Lista de autores
          content:
              application/json:
                schema:
                  type: array
                  items:
                    $ref: '#/components/schemas/Author'
    post:
      summary: Crear un nuevo autor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorInput'
            example:
              name: "Milan Kundera"
              email: "milan@kundera.cz"
              bio: "Escritor checo, autor de 'La insoportable levedad del ser"
      responses:
        '201':
          description: Autor creado.
        '400':
          $ref: '#/components/responses/BadRequest'
  /authors/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      summary: Obtener detalles de un autor  
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Actualizar información de un autor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorInput'
      responses:
        '200':
          description: Autor actualizado.
        '404':
          $ref: '#/components/responses/NotFound'      
    delete:
      summary: Eliminar un autor y sus posts/comentarios asociados
      responses:
        '204':
          description: Autor eliminado correctamente.
        '404':
          $ref: '#/components/responses/NotFound'  
  /posts:
    get:
      summary: Listar todos los posts
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
    post:
      summary: Crear un nuevo post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostInput'
            example:
              title: "La insoportable levedad del ser"
              content: Un ensayo sobre el peso de las desiciones.
              author_id: 1
              published: true
      responses:
        '201':
          description: Post creado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'
  /posts/{id}:
    parameters:
      -$ref: '#/components/parameters/Id'
    get:
      summary: Obtener post por ID
      responses:
        '200':
          description: Detalles del post.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Actualizar un post específico
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostInput'
      responses:
        '200':
          description: Post actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Eliminar un post específico
      responses:
        '204':
          description: Post eliminado correctamente.
        '404':
          $ref: '#/components/responses/NotFound'
  /posts/author/{authorId}:
    parameters:
      - name: authorId
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Listar posts de un autor específico
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostWithAuthor'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Actualizar post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostInput'
      responses:
        '200':
          description: Post actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /comments:
    post:
      summary: Crear un comentario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentInput'
      responses:
        '201':
          description: Comentario creado.
  /comments/post/{postId}:
    parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Ver comentarios de un post
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentWithAuthor'
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
  responses:
    BadRequest:
      description: Solicitud inválida (faltan datos o estan mal formados).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: El recurso solicitado no existe.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas: 
    Author:
      type: object
      properties:
        id: {type: integer}
        name: {type: string}
        email: {type: string, format: email}
        bio: {type: string, nullable: true}
        created_at: {type: string, format: date-time}
    AuthorInput:
      type: object
      required: [name, email]
      properties:
        name: {type: string}
        email: {type: string}
        bio: {type: string}
    Post:
      type: object
      properties:
        title: {type: string}
        content: {type: string}
        author_id: {type: integer}
        published: {type: boolean} 
    PostInput:
      type: object
      required: [title, content, author_id]
      properties:
        title: {type: string}
        content: {type: string}
        author_id: {type: integer}
        published: {type: boolean}
    PostWithAuthor:
      allOf:
        - $ref: '#/components/schemas/Post'
        - type: object
          properties:
            author_name: {type: string}
    Comment:
      type: object
      properties:
        id: {type: integer}
        content: {type: string}
        author_id: {type: integer}
        post_id: {type: integer}
    CommentInput:
      type: object
      required: [content, author_id, post_id]
      properties:
        content: {type: string}
        author_id: {type: integer}
        post_id: {type: integer}
    CommentWithAuthor:
      allOf:
        - $ref: '#/components/schemas/Comment'
        - type: object
          properties:
            author_name: {type: string}
    Error:
      type: object
      properties:
        message: {type: string}